Move renderer struct from library file to new renderer file
Took 11 minutes
